Updated_schema_freq.py
'The List[float] str not supported' error turns out to be a mapping problem. so I changed 'List' - list to 'List[float] - list, 'from the mapping.
Then, when you try to validate, a ''key': Line' error arises. I have read that this will be solved by defining a schema dictionary that contains the 'Line' and 'Cont'. So I defined a schema dictionary to solve the key error.

the code works for me. check it out and let me know.
And we might not need a  Regex.

(there are some more changes too)