Updated_schema_freq.py
The ''the List[float] str not supported'' error turns out to be a mapping problem. so I changed the mapping from 'list' - list, to 'List[float]' - list.
When you run the code after this change, a 'key': 'Line', error arises. This could be solved by defining a schema dictionary that contains Line and Cont.
So I defined a schema. I did not use the regex.